youtube_telegram_bot/requirements.txt
python-telegram-bot>=21.0
yt-dlp==2023.12.30
python-dotenv==1.0.0